+++ urbanairship-automation/src/test/java/com/urbanairship/TestApplication.java
@@ -8,6 +8,7 @@ import android.content.pm.ProviderInfo;
 import com.urbanairship.analytics.Analytics;
 import com.urbanairship.channel.AirshipChannel;
 import com.urbanairship.js.Whitelist;
+import com.urbanairship.locale.LocaleManager;
 import com.urbanairship.modules.location.AirshipLocationClient;
 import com.urbanairship.push.PushManager;
 import com.urbanairship.util.PlatformUtils;
@@ -48,8 +49,9 @@ public class TestApplication extends Application implements TestLifecycleApplica
         UAirship.sharedAirship = new UAirship(airshipConfigOptions);
         UAirship.sharedAirship.preferenceDataStore = preferenceDataStore;
         UAirship.sharedAirship.runtimeConfig = testRuntimeConfig;
-        UAirship.sharedAirship.channel = new AirshipChannel(this, preferenceDataStore, UAirship.sharedAirship.runtimeConfig);
-        UAirship.sharedAirship.analytics = new Analytics(this, preferenceDataStore, testRuntimeConfig, UAirship.sharedAirship.channel);
+        UAirship.sharedAirship.localeManager = new LocaleManager(this, preferenceDataStore);
+        UAirship.sharedAirship.channel = new AirshipChannel(this, preferenceDataStore, UAirship.sharedAirship.runtimeConfig, UAirship.sharedAirship.localeManager);
+        UAirship.sharedAirship.analytics = new Analytics(this, preferenceDataStore, testRuntimeConfig, UAirship.sharedAirship.channel,  UAirship.sharedAirship.localeManager);
         UAirship.sharedAirship.pushManager = new PushManager(this, preferenceDataStore, airshipConfigOptions, null, UAirship.sharedAirship.channel, UAirship.sharedAirship.analytics);
         UAirship.sharedAirship.whitelist = Whitelist.createDefaultWhitelist(airshipConfigOptions);